refactor(GameView): delete GameView class, add InputView, OutputView class
